Changed the data type of the  primary key of the  "Lessons" entity to byte (instead of int)

+++ Project/Core/Domains/Lesson.cs
@@ -2,10 +2,10 @@
 {
     public class Lesson
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
         public string Title { get; set; }
         public string Body { get; set; }
-        public byte Index { get; set; }
+        public int Index { get; set; }
         public Course Course { get; set; }
         public int CourseId { get; set; }
     }
